Extract file importation logic from view model to manger
